mkfs can improperly generate an error when the data subvolume stripe unit is larger than 256kb, which is larger than the maximum log stripe unit size.
So, only set and check the log stripe unit on version 2 internal logs, and
default the stripe unit size to 32kb if the data volume's stripe unit
size is too big.
Only set and check the log stripe unit on version 2 internal logs, and
default the stripe unit size to 32kb if the data volume's stripe unit
size is too big.