xfs_repair doesn't handle unwritten real-time extents properly.  When the real-time subvolume extent size is larger than a filesystem block, it is possible to have a real-time extent that is part written and part unwritten.
Internally, xfs_repair tracks real-time extent use on a per-extent
basis.  Thus, a file that has a written and unwritten extent falling
on the same extent.

This fix addresses this problem specificly by noting when the file
extent is unwritten and covers a partial real-time extent.  Then,
instead of checking that it is unused, it checks that it is used.