This  mod implements a new algorithm for phase7 which makes use of the recently introduced prefetch framework.
Instead of using the transaction framework one inode at a
time, we read large chunks of inodes off the disk in chunks
and compare the nlink count.
This  mod implements a new algorithm for phase7 which
makes use of the recently introduced prefetch framework.

Instead of using the transaction framework one inode at a 
time, we read large chunks of inodes off the disk in chunks 
and compare the nlink count.