xfs_check: fix exit codes
xfs_check returns success when the superblock magic is invalid - this makes it
return the serious error exit code (3) for this case, and the invalid arguments
exit code (1) if the arguments to xfs_check were wrong.

A glance at the code suggests that other xfs_check failures appear to return
the correct error codes.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Cc: Michael Sinz <michael.sinz@sinz.org>