xfs_repair: allow filesystems with a single allocation group
Currently xfs_repair bails out on a filesystem with just a single AG.
But that's a perfectly valid configureation, so we should allow it.

Because we could cause harm only allow it when using the force_geometry
suboption.  But we nice enough to tell the user about it when he needs it.

Also make sure to take the internal log into account when guestimating the
first inode cluster offset.

I'll also cook up a testcase for repair on single AG filesystems.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reported-By: Arkadiusz Miskiewicz <arekm@maven.pl>
Reviewed-by: Eric Sandeen <sandeen@sandeen.net>