fix parallel builds
I noticed that xfsprogs-3.0.0 sometimes fails to build in parallel with an
error about fsr not being able to find libhandle.  Looking at the top level
Makefile shows the obvious missing depend (fsr: libhandle).  Playing around
a bit found a few more not so obvious missing depends on the include subdir
so I fixed that too.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reviewed-by: Eric Sandeen <sandeen@sandeen.net>