keep autoconf-generated files over make distclean
We do want to ship the autoconf-generated files over make distclean and
pick them up in Makepkgs so that a user compiling the program doesn't
require autoconf.

For that split up make distclean from make realclean and exclude the
files we want to keep form the former.

To make this easier to maintain adher to stricter rules of use for
CONFIGURE and LDIRT.  CONFIGURE now includes all generated files from
autoconf (or for it like our copied install-sh) and gets only removed
on make realclean, but added to LSRCFILES so that Makepkgs can pick it
up.  Everything else than needs to be removed on a realclean gets
assigned to LDIRT.

Also split up the configure rule into one to generate the configure
script and one to run it.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Mike Frysinger <vapier@gentoo.org>