xfsprogs: fix the -V option for most shell scripts
For most of the shellscripts wrapping xfs_db or xfs_growfs the -V option
is not implemented correctly.  If we just add -V to the options for the
underlying binaries we will never actually call it because the mandatory
device / mount point argument is missing.  Instead just call the underlying
command with -V directly and exit early.  This is what xfs_bmap.sh,
xfs_mkfile and xfs_metadump.sh are already doing.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reported-by: Michael Monnerie <michael.monnerie@is.it-management.at>
Reviewed-by: Eric Sandeen <sandeen@sandeen.net>