simplify and unify source package generation
Currently the build/ directory can create rpm, debian and source / binary
tar packages.  The RPM generation is not used as all distributions prefer
their own spec files, and the binary tarball not used at all as it's
a not very useful format.  Reimplement the generation of the source
tarballs to use the source-link method used for the debian packages and
get rid of the whole old package generation machinery.  Also fix a small
bug in the link-based source directory creation which was not including
the .pot file for gettext.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Nathan Scott <nathans@debian.org>