xfs_db: modify bad_features2 when modifying features2
The "attr1" command in xfs_db, for example, only modifies the features2
field; when mounted, the kernel will find a mismatch between features2
and bad_features2, and attr2 gets turned back on.

I think the simplest fix is to modify do_version to modify both fields,
but not if there is an existing mismatch that should be investigated
first.

Any mismatch can be fixed up by writing directly to the superblock
fields.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>
Reviewed-by: Christoph Hellwig <hch@lst.de>