xfs_fsr: don't create ag temp subdirs with mode 0777
This addresses Debian bug #559490 - xfsdump: fsr creates files mode 666

I think that this mode is left over from when xfs_fsr used to
fork into the uid of the file's owner, and so needed somewhere
it was guaranteed to be able to write.

This behavior was removed in commit d51b892411c8d33374a02e20c5888df280811549
(in the xfsdump tree, before xfs_fsr got moved) and so these
wide-open permissions should no longer be needed.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>
Reviewed-by: Nathan Scott <nscott@aconex.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>