libhandle: always use a good path for by-handle ioctls
We can't open symbolic links to perform the by-handle XFS ioctls, and
while we can open special files they won't end up calling into the
XFS ioctl method.  So before calling into the handle ioctls generate
a fspath that always points to a regular file or directory that we can
call the ioctl on.

Signed-off-by: Bill Kendall <wkendall@sgi.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>