xfsprogs: metadump: ensure dup table always has entry for obfuscated name
We need to ensure the nametable has a copy of all the names in a
directory (or attribute fork) in order to avoid creating duplicate
entries when obfuscating names.  Currently there is an (unlikely)
case where the name is passed back without such an entry being
created.  Reorder things so that won't happen.

Signed-off-by: Alex Elder <aelder@sgi.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>