xfsprogs: avoid dot-directories when configuring
The "find" command used in the configure script to find localized
files searches through directories (including .git and .pc) that
really should be ignored.  Change it so it skips over these
directories.

I think it's reasonable to assume any such "dot directory" should be
ignored, so this change skips any directory at the top level whose
name begins with ".".

Note that I found an odd anomaly in "find".  If you do not supply
the "-print" argument, the pruned directory names show up in the
output.  Supplying "-print" does not include them (and that's what
we want).

Signed-off-by: Alex Elder <aelder@sgi.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>